+++ server/src/models/credential.model.js
@@ -16,8 +16,10 @@ var Credential=function(credential){
     this.host = credential.host;
     this.port = credential.port;
     this.user = credential.user;
+    this.secure = (credential.secure)?1:0;
     this.password = encrypt(credential.password);
     this.description = credential.description;
+    this.db_type = credential.db_type;
 };
 
 Credential.create = function (record) {
@@ -39,7 +41,7 @@ Credential.delete = function(id){
 };
 Credential.findAll = function () {
     logger.info("Finding all credentials")
-    return mysql.do("SELECT id,name,user,host,port,description FROM AnsibleForms.`credentials`;")
+    return mysql.do("SELECT id,name,user,host,port,description,secure,db_type FROM AnsibleForms.`credentials`;")
 };
 Credential.findById = function (id) {
     logger.info(`Finding credential ${id}`)
@@ -79,7 +81,7 @@ Credential.findByName = function (name) {
   logger.debug(`Finding credential ${name}`)
   var cred = cache.get(name)
   if(cred==undefined){
-    return mysql.do("SELECT host,port,name,user,password FROM AnsibleForms.`credentials` WHERE name=?;",name)
+    return mysql.do("SELECT host,port,name,user,password,secure,db_type FROM AnsibleForms.`credentials` WHERE name=?;",name)
     .then((res)=>{
       if(res.length>0){
         res[0].multipleStatements = true